Move RMSNorm with the model
